+++ source/extensions/omni.isaac.orbit_assets/docs/CHANGELOG.rst
@@ -1,6 +1,15 @@
 Changelog
 ---------
 
+0.1.1 (2024-03-11)
+~~~~~~~~~~~~~~~~~~
+
+Added
+^^^^^
+
+* Added configurations for allegro and shadow hand assets.
+
+
 0.1.0 (2023-12-20)
 ~~~~~~~~~~~~~~~~~~